checkout-single-subscription/requirements.txt
alembic==1.13.1
asgiref==3.7.2
attrs==23.1.0
blinker==1.6.3
cachetools==5.3.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dnspython==2.6.1
dominate==2.9.1
email-validator==2.1.0.post1
Faker==21.0.0
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flask_httpauth==4.8.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.2
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.2
MarkupSafe==2.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
stripe==7.0.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
visitor==0.1.3
Werkzeug==3.0.0
wsproto==1.2.0
WTForms==3.1.2
